+++ Banjo/Source/Utility/Math/Vector3.h
@@ -5,7 +5,7 @@ class Vector3
 {
 public:
     Vector3();
-    Vector3(int x,  int y,  int z);
+    Vector3(float x,  float y,  float z);
 
     Vector3 operator+(Vector3 const& vector) const;
     Vector3 operator +=(Vector3 const& vector);
